added jobs api

+++ OnlineJobPortal-Backend/src/main/java/com/example/OnlineJobPortal/repo/JobsRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface JobsRepo extends JpaRepository<Jobs, Long> {
+public interface JobsRepo extends JpaRepository<Jobs, Integer> {
 }
